requirements.txt
browser-use==0.11.4
langchain-google-vertexai==3.2.1
playwright==1.57.0
python-dotenv==1.2.1
pyyaml==6.0.3
openpyxl==3.1.2
